+++ util.h
@@ -6,16 +6,16 @@
 using namespace std;
 
 typedef struct Point{
-    double x, y;
+    float x, y;
 }Point;
 
 void read_file(char *filename, vector <Point> &points);
 
-double kmeans_cpu(Point* points, Point* means,int* labels,double* dist, int iter, int n, int k);
-double kmeans_gpu(Point* points, Point* cpupoints, Point* means,int* labels,double* dist,double* cpudist, int iter, int n, int k);
-double kmeans_cpu_ineq(Point* points,Point* means, int* labels,double* icd,int* rid,int iter,int n,int k);
-double kmeans_gpu_ineq(Point* points, Point* means, int* labels, double* icd, int* rid, int iter, int n, int k);
-vector <vector <double>> gmix_gpu(vector <Point> points, int iter);
+float kmeans_cpu(Point* points, Point* means,int* labels,float* dist, int iter, int n, int k);
+float kmeans_gpu(Point* points, Point* cpupoints, Point* means,int* labels,float* dist,float* cpudist, int iter, int n, int k);
+float kmeans_cpu_ineq(Point* points,Point* means, int* labels,float* icd,int* rid,int iter,int n,int k);
+float kmeans_gpu_ineq(Point* points, Point* means, int* labels, float* icd, int* rid, int iter, int n, int k);
+vector <vector <float>> gmix_gpu(vector <Point> points, int iter);
 
-vector <vector <double>> gmix_cpu(vector <Point> points, int iter);
+vector <vector <float>> gmix_cpu(vector <Point> points, int iter);
